fix tests for sev and grpc
