+++ internal/server/grpc/grpc_test.go
@@ -31,6 +31,28 @@ const bufSize = 1024 * 1024
 
 var lis *bufconn.Listener
 
+type DummyRWC struct{}
+
+// Read fills p with byte(len(p)) and returns len(p).
+func (l *DummyRWC) Read(p []byte) (int, error) {
+	n := len(p)
+	// Fill each byte in p with the value of n as a byte.
+	for i := range p {
+		p[i] = byte(n)
+	}
+	return n, nil
+}
+
+// Write simply returns len(p) indicating that all bytes were written.
+func (l *DummyRWC) Write(p []byte) (int, error) {
+	// In this simple implementation, we ignore the data.
+	return len(p), nil
+}
+
+func (l *DummyRWC) Close() error {
+	return nil
+}
+
 func init() {
 	lis = bufconn.Listen(bufSize)
 }
@@ -260,7 +282,7 @@ func (b *ThreadSafeBuffer) String() string {
 }
 
 func TestServerInitializationAndStartup(t *testing.T) {
-	vtpm.ExternalTPM = &vtpm.DummyRWC{}
+	vtpm.ExternalTPM = &DummyRWC{}
 
 	testCases := []struct {
 		name          string
