+++ scripts/attestation_policy/pcr_values.json
@@ -24,4 +24,3 @@
     }
   }
 }
-
